Created playRound and game functions.
playRound function plays a round of rock paper
scissors with the computer.

game function uses for loop to run playRound
function for 5 times and keeps count of who won
which round and tells the overall winner.

error: getUserValue function needs a bit of dibugging as
the game does not work as intented if we get value from the user.